backend/pyproject.toml
[project]
name = "ui-auto-backend"
version = "0.1.0"
description = "Backend for UI Automation Platform"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn>=0.27.0",
    "sqlalchemy>=2.0.25",
    "asyncpg>=0.29.0",
    "alembic>=1.13.1",
    "pydantic-settings>=2.1.0",
    "playwright>=1.41.0",
    "python-multipart>=0.0.6",
    "pyjwt>=2.8.0",
    "passlib[bcrypt]>=1.7.4",
    "bcrypt==4.0.1",
    "redis>=5.0.1",
    "allure-python-commons>=2.13.2",
    "greenlet>=3.0.3",
    "email-validator>=2.3.0",
    "celery[redis]>=5.3.6",
    "nest-asyncio>=1.6.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]
